Fix: fix test on linux platform
